resume-text-converter/requirements.txt
azure-functions
PyPDF2
python-docx
